Remove deprecated symbols
