+++ src/PersonalId.php
@@ -18,32 +18,32 @@ class PersonalId extends AbstractId
      * @var string[] Map of county number high limit to county identifier
      */
     private static $birthCountyMap = [
-        13 => Id::COUNTY_STOCKHOLM,
-        15 => Id::COUNTY_UPPSALA,
-        18 => Id::COUNTY_SODERMANLAND,
-        23 => Id::COUNTY_OSTERGOTLAND,
-        26 => Id::COUNTY_JONKOPING,
-        28 => Id::COUNTY_KRONOBERG,
-        31 => Id::COUNTY_KALMAR,
-        32 => Id::COUNTY_GOTLAND,
-        34 => Id::COUNTY_BLEKINGE,
-        38 => Id::COUNTY_KRISTIANSTAD,
-        45 => Id::COUNTY_MALMOHUS,
-        47 => Id::COUNTY_HALLAND,
-        54 => Id::COUNTY_GOTEBORG_BOUHUS,
-        58 => Id::COUNTY_ALVSBORG,
-        61 => Id::COUNTY_SKARABORG,
-        64 => Id::COUNTY_VARMLAND,
-        65 => Id::COUNTY_UNDEFINED,
-        68 => Id::COUNTY_OREBRO,
-        70 => Id::COUNTY_VASTMANLAND,
-        73 => Id::COUNTY_KOPPARBERG,
-        74 => Id::COUNTY_UNDEFINED,
-        77 => Id::COUNTY_GAVLEBORG,
-        81 => Id::COUNTY_VASTERNORRLAND,
-        84 => Id::COUNTY_JAMTLAND,
-        88 => Id::COUNTY_VASTERBOTTEN,
-        92 => Id::COUNTY_NORRBOTTEN
+        13 => IdInterface::COUNTY_STOCKHOLM,
+        15 => IdInterface::COUNTY_UPPSALA,
+        18 => IdInterface::COUNTY_SODERMANLAND,
+        23 => IdInterface::COUNTY_OSTERGOTLAND,
+        26 => IdInterface::COUNTY_JONKOPING,
+        28 => IdInterface::COUNTY_KRONOBERG,
+        31 => IdInterface::COUNTY_KALMAR,
+        32 => IdInterface::COUNTY_GOTLAND,
+        34 => IdInterface::COUNTY_BLEKINGE,
+        38 => IdInterface::COUNTY_KRISTIANSTAD,
+        45 => IdInterface::COUNTY_MALMOHUS,
+        47 => IdInterface::COUNTY_HALLAND,
+        54 => IdInterface::COUNTY_GOTEBORG_BOUHUS,
+        58 => IdInterface::COUNTY_ALVSBORG,
+        61 => IdInterface::COUNTY_SKARABORG,
+        64 => IdInterface::COUNTY_VARMLAND,
+        65 => IdInterface::COUNTY_UNDEFINED,
+        68 => IdInterface::COUNTY_OREBRO,
+        70 => IdInterface::COUNTY_VASTMANLAND,
+        73 => IdInterface::COUNTY_KOPPARBERG,
+        74 => IdInterface::COUNTY_UNDEFINED,
+        77 => IdInterface::COUNTY_GAVLEBORG,
+        81 => IdInterface::COUNTY_VASTERNORRLAND,
+        84 => IdInterface::COUNTY_JAMTLAND,
+        88 => IdInterface::COUNTY_VASTERBOTTEN,
+        92 => IdInterface::COUNTY_NORRBOTTEN
     ];
 
     /**
@@ -121,6 +121,6 @@ class PersonalId extends AbstractId
             }
         }
 
-        return Id::COUNTY_UNDEFINED;
+        return IdInterface::COUNTY_UNDEFINED;
     }
 }
