+++ tests/NullIdTest.php
@@ -55,7 +55,7 @@ class NullIdTest extends TestCase
         );
     }
 
-    public function testGetDate()
+    public function testGetBirthDate()
     {
         $this->expectException('byrokrat\id\Exception\DateNotSupportedException');
         (new NullId)->getBirthDate();
@@ -64,7 +64,7 @@ class NullIdTest extends TestCase
     public function testGetSex()
     {
         $nullId = new NullId();
-        $this->assertEquals(Id::SEX_UNDEFINED, $nullId->getSex());
+        $this->assertEquals(IdInterface::SEX_UNDEFINED, $nullId->getSex());
         $this->assertTrue($nullId->isSexUndefined());
         $this->assertFalse($nullId->isMale());
         $this->assertFalse($nullId->isFemale());
@@ -73,7 +73,7 @@ class NullIdTest extends TestCase
     public function testGetLegalForm()
     {
         $nullId = new NullId();
-        $this->assertEquals(Id::LEGAL_FORM_UNDEFINED, $nullId->getLegalForm());
+        $this->assertEquals(IdInterface::LEGAL_FORM_UNDEFINED, $nullId->getLegalForm());
         $this->assertTrue($nullId->isLegalFormUndefined());
         $this->assertFalse($nullId->isStateOrParish());
         $this->assertFalse($nullId->isIncorporated());
@@ -86,7 +86,7 @@ class NullIdTest extends TestCase
     public function testGetBirthCounty()
     {
         $this->assertEquals(
-            Id::COUNTY_UNDEFINED,
+            IdInterface::COUNTY_UNDEFINED,
             (new NullId)->getBirthCounty()
         );
     }
